refactor: rename logger
